+++ src/components/AlgoButtons.tsx
@@ -1,0 +1,9 @@
+export default function AlgoButtons() {
+    return(
+        <div>
+            <button>
+                
+            </button>
+        </div>
+    )
+}
